+++ schemas/credential.js
@@ -2,9 +2,11 @@
  * Copyright (c) 2015 Digital Bazaar, Inc. All rights reserved.
  */
 var bedrock = require('bedrock');
+var constants = bedrock.config.constants;
 
 var schemas = {};
 schemas.identifier = require('./identifier');
+schemas.jsonldContext = require('./jsonldContext');
 schemas.url = require('./url');
 schemas.w3cDateTime = require('./w3cDateTime');
 
@@ -13,7 +15,12 @@ var schema = {
   type: 'object',
   title: 'Credential',
   properties: {
-    '@context': {required: true},
+    '@context': schemas.jsonldContext(),
+    // FIXME: improve credential context check
+    //'@context': schemas.jsonldContext([
+    //  constants.IDENTITY_CONTEXT_V1_URL,
+    //  constants.CREDENTIALS_CONTEXT_V1_URL
+    //]),
     issuer: schemas.identifier({required: true}),
     issued: schemas.w3cDateTime({required: true}),
     claim: {